+++ scripts/src/local/commands/test.ts
@@ -11,7 +11,7 @@ import { existsSync, readdirSync, readFileSync } from "node:fs";
 import { join } from "node:path";
 import { homedir } from "node:os";
 import { execFileSync } from "node:child_process";
-import { parseSkillContent } from "../../shared/parse-skill.js";
+import { parseSkillContent } from "../../shared/skill-helper.js";
 
 const TIMEOUT_MS = 120_000;
 const MARKETPLACE_NAME = "github-copilot-for-azure";
@@ -62,7 +62,7 @@ function checkPluginConfig(expectedCachePath: string): TestResult {
 
   const marketplace = config.marketplaces?.[MARKETPLACE_NAME];
   if (!marketplace || marketplace.source?.source !== "github" ||
-      marketplace.source?.repo !== "microsoft/github-copilot-for-azure") {
+    marketplace.source?.repo !== "microsoft/github-copilot-for-azure") {
     return { name: "Plugin config", passed: false, detail: `Marketplace "${MARKETPLACE_NAME}" not configured correctly` };
   }
