+++ scripts/src/shared/string-helpers.ts
