+++ scripts/src/tokens/commands/types.ts
@@ -167,8 +167,8 @@ export function matchesPattern(filePath: string, pattern: string): boolean {
     return normalizedPath.endsWith("/" + pattern) || normalizedPath === pattern;
   }
 
-  // @ts-expect-error Using named import satisfies TypeScript's static analysis but the code won't work at runtime. Using default import works at runtime but fails the static analysis.
-  return isMatch(normalizedPath, pattern);
+  const matches = isMatch([normalizedPath], pattern);
+  return matches.length > 0
 }
 
 /**
